+++ src/pyscipopt/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '3.1.5'
+__version__ = '3.2.0'
 
 # required for Python 3.8 on Windows
 import os